refactor/#28: CQRS 패턴 적용
